+++ sdk/deposit/log/index.js
@@ -1,6 +1,7 @@
 const log = require('./log.js');
 
 exports.get = log.get;
+exports.pdf = log.pdf;
 exports.query = log.query;
 exports.page = log.page;
 exports.Log = log.Log;