+++ index.js
@@ -1,4 +1,4 @@
-exports.version = '2.31.0';
+exports.version = '2.32.0';
 
 exports.cache = {};
 exports.user = {};