+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Tax as a valid payment request
 
 ## [2.31.0] - 2025-03-27
 ### Added